# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * USTAV Invitation Card Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  mobile: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface CustomerRegistration {
  name: string;
  email: string;
  mobile: string;
  /** @minLength 6 */
  password: string;
}

export interface AdminRegistration {
  name: string;
  email: string;
  mobile: string;
  /** @minLength 6 */
  password: string;
}

export type LoginInputRole =
  (typeof LoginInputRole)[keyof typeof LoginInputRole];

export const LoginInputRole = {
  customer: "customer",
  admin: "admin",
} as const;

export interface LoginInput {
  email: string;
  password: string;
  role: LoginInputRole;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @nullable */
  iconUrl: string | null;
  templateCount: number;
  createdAt: string;
}

export interface Template {
  id: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  name: string;
  description: string;
  imageUrl: string;
  price: number;
  isFeatured: boolean;
  createdAt: string;
}

export interface CategoryWithTemplates {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @nullable */
  iconUrl: string | null;
  templateCount: number;
  createdAt: string;
  templates: Template[];
}

export interface CategoryInput {
  name: string;
  description: string;
  /** @nullable */
  iconUrl?: string | null;
}

export interface TemplateInput {
  categoryId: number;
  name: string;
  description: string;
  imageUrl: string;
  price: number;
  isFeatured?: boolean;
}

export interface TemplateUpdate {
  categoryId?: number;
  name?: string;
  description?: string;
  imageUrl?: string;
  price?: number;
  isFeatured?: boolean;
}

export interface UploadResponse {
  url: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface InvitationCustomization {
  hostName?: string;
  eventDate?: string;
  eventTime?: string;
  venue?: string;
  rsvpDetails?: string;
  customMessage?: string;
  guestName?: string;
}

export interface Order {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  templateId: number;
  /** @nullable */
  templateName?: string | null;
  /** @nullable */
  templateImageUrl?: string | null;
  status: OrderStatus;
  totalPrice: number;
  customization: InvitationCustomization;
  createdAt: string;
}

export interface OrderInput {
  templateId: number;
  customization: InvitationCustomization;
}

export type OrderUpdateStatus =
  (typeof OrderUpdateStatus)[keyof typeof OrderUpdateStatus];

export const OrderUpdateStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface OrderUpdate {
  status?: OrderUpdateStatus;
  customization?: InvitationCustomization;
}

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  completedOrders: number;
  totalRevenue: number;
  totalTemplates: number;
  totalCategories: number;
  totalCustomers: number;
}

export type ListTemplatesParams = {
  categoryId?: number;
  featured?: boolean;
};