# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: USTAV Invitation Card Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register/customer:
    post:
      operationId: registerCustomer
      tags: [auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegistration"
      responses:
        "201":
          description: Customer registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register/admin:
    post:
      operationId: registerAdmin
      tags: [auth]
      summary: Register a new admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRegistration"
      responses:
        "201":
          description: Admin registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (customer or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all event categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new event category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a single category with its templates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category with templates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithTemplates"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (customer sees own orders, admin sees all)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order for a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status (admin) or customization (customer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    User:
      type: object
      required: [id, name, email, mobile, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        role:
          type: string
          enum: [customer, admin]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    CustomerRegistration:
      type: object
      required: [name, email, mobile, password]
      properties:
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        password:
          type: string
          minLength: 6
    AdminRegistration:
      type: object
      required: [name, email, mobile, password]
      properties:
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, admin]
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    Category:
      type: object
      required: [id, name, slug, description, iconUrl, templateCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        iconUrl:
          type: ["string", "null"]
        templateCount:
          type: integer
        createdAt:
          type: string
    CategoryWithTemplates:
      type: object
      required: [id, name, slug, description, iconUrl, templateCount, createdAt, templates]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        iconUrl:
          type: ["string", "null"]
        templateCount:
          type: integer
        createdAt:
          type: string
        templates:
          type: array
          items:
            $ref: "#/components/schemas/Template"
    CategoryInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        iconUrl:
          type: ["string", "null"]
    Template:
      type: object
      required: [id, categoryId, name, description, imageUrl, price, isFeatured, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        isFeatured:
          type: boolean
        createdAt:
          type: string
...
    Order:
      type: object
      required: [id, customerId, templateId, status, totalPrice, customization, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        templateId:
          type: integer
        templateName:
          type: ["string", "null"]
        templateImageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
        totalPrice:
          type: number
        customization:
          $ref: "#/components/schemas/InvitationCustomization"
        createdAt:
          type: string
    InvitationCustomization:
      type: object
      properties:
        hostName:
          type: string
        eventDate:
          type: string
        eventTime:
          type: string
        venue:
          type: string
        rsvpDetails:
          type: string
        customMessage:
          type: string
        guestName:
          type: string
    OrderInput:
      type: object
      required: [templateId, customization]
      properties:
        templateId:
          type: integer
        customization:
          $ref: "#/components/schemas/InvitationCustomization"
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
        customization:
          $ref: "#/components/schemas/InvitationCustomization"
...