# lib/api-zod/src/generated/types/adminRegistration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * USTAV Invitation Card Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminRegistration {
  name: string;
  email: string;
  mobile: string;
  /** @minLength 6 */
  password: string;
}