# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * USTAV Invitation Card Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  name: string;
  description: string;
  /** @nullable */
  iconUrl?: string | null;
}