# lib/api-zod/src/generated/types/categoryWithTemplates.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * USTAV Invitation Card Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Template } from "./template";

export interface CategoryWithTemplates {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @nullable */
  iconUrl: string | null;
  templateCount: number;
  createdAt: string;
  templates: Template[];
}