# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * USTAV Invitation Card Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InvitationCustomization } from "./invitationCustomization";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  templateId: number;
  /** @nullable */
  templateName?: string | null;
  /** @nullable */
  templateImageUrl?: string | null;
  status: OrderStatus;
  totalPrice: number;
  customization: InvitationCustomization;
  createdAt: string;
}