# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * USTAV Invitation Card Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InvitationCustomization } from "./invitationCustomization";

export interface OrderInput {
  templateId: number;
  customization: InvitationCustomization;
}